Simplify $ABBR record grammar and get rid of WS_ALL token definition
