+++ src/pharmpy/internals/parse/__init__.py
@@ -1,3 +1,4 @@
 from .generic import AttrToken, AttrTree, GenericParser, NoSuchRuleException
+from .ignored import with_ignored_tokens
 
-__all__ = ('AttrTree', 'AttrToken', 'NoSuchRuleException', 'GenericParser')
+__all__ = ('AttrTree', 'AttrToken', 'NoSuchRuleException', 'GenericParser', 'with_ignored_tokens')
