+++ src/pharmpy/plugins/nonmem/records/grammars/definitions.lark
@@ -1,5 +1,3 @@
-WS_ALL: /\s+/
-
 WS: /[ \x00\t]+/
 NEWLINE: /\r?\n/
 COMMENT: /;[^\r\n]*/
@@ -14,7 +12,7 @@ SIGN: "+"
 
 EXP: ("E" | "e") SIGNED_INT
 FLOAT: INT EXP
-     | DECIMAL [EXP]
+     | DECIMAL EXP?
 SIGNED_FLOAT: SIGN? FLOAT
 
 NUMBER: FLOAT | INT
