+++ src/pharmpy/plugins/nonmem/records/parsers.py
@@ -1,10 +1,8 @@
 from pathlib import Path
-from typing import Iterator, List, Tuple, Union
 
-from lark import Lark, Token, Transformer, Tree, Visitor
-from lark.tree import Meta
+from lark import Lark, Tree, Visitor
 
-from pharmpy.internals.parse import GenericParser
+from pharmpy.internals.parse import GenericParser, with_ignored_tokens
 
 grammar_root = Path(__file__).resolve().parent / 'grammars'
 
@@ -23,6 +21,10 @@ class RecordParser(GenericParser):
 @install_grammar
 class AbbreviatedRecordParser(RecordParser):
     grammar_filename = 'abbreviated_record.lark'
+    grammar_options = dict(
+        propagate_positions=True,
+    )
+    post_process = [with_ignored_tokens]
 
 
 @install_grammar
@@ -50,108 +52,6 @@ class InitOrLow(Visitor):
                 tree.data = 'low' if is_low else 'init'
 
 
-WS = {' ', '\x00', '\t'}
-
-
-def _tokenize_ignored_characters(s: str, i: int, j: int):
-    # TODO propagate line/column information
-    head = i
-    while i < j:
-        first = s[head]
-        head += 1
-        if first in WS:
-            while head < j and s[head] in WS:
-                head += 1
-
-            yield Token('WS', s[i:head], start_pos=i, end_pos=head)
-
-        elif first == ';':
-            while head < j and s[head] != '\n' and s[head] != '\r':
-                head += 1
-
-            yield Token('COMMENT', s[i:head], start_pos=i, end_pos=head)
-
-        elif first == '\r':
-            assert head < j and s[head] == '\n'
-            head += 1
-
-            yield Token('NEWLINE', s[i:head], start_pos=i, end_pos=head)
-
-        else:
-            assert first == '\n'
-            yield Token('NEWLINE', s[i:head], start_pos=i, end_pos=head)
-
-        i = head
-
-
-def _item_range(x: Union[Tree, Token]) -> Tuple[int, int]:
-    if isinstance(x, Tree):
-        i = x.meta.start_pos
-        j = x.meta.end_pos
-    else:
-        i = x.start_pos
-        j = x.end_pos
-
-    assert isinstance(i, int)
-    assert isinstance(j, int)
-    return (i, j)
-
-
-def _interleave_ignored(source: str, it: Iterator[Union[Tree, Token]]):
-    x = next(it)
-    yield x
-
-    i = _item_range(x)[1]
-
-    x = next(it)
-    while True:
-        j, k = _item_range(x)
-
-        if i < j:
-            yield from _tokenize_ignored_characters(source, i, j)
-
-        yield x
-
-        try:
-            x = next(it)
-        except StopIteration:
-            break
-
-        i = k
-
-
-def interleave_ignored(source: str, children: List[Union[Tree, Token]]):
-    if len(children) < 2:
-        return children
-    else:
-        return list(_interleave_ignored(source, iter(children)))
-
-
-class InterleaveIgnored(Transformer):
-    def __init__(self, source: str):
-        self.source = source
-
-    def __default__(self, data, children, meta):
-        new_children = interleave_ignored(self.source, children)
-        return Tree(data, new_children, meta)
-
-
-def insert_ignored(source, tree):
-    new_tree = InterleaveIgnored(source).transform(tree)
-    final_meta = Meta()
-    # TODO propagate line/column information
-    final_meta.start_pos = 0
-    final_meta.end_pos = len(source)
-    final_meta.empty = False
-    final_children = (
-        list(_tokenize_ignored_characters(source, 0, new_tree.meta.start_pos))
-        + new_tree.children
-        + list(_tokenize_ignored_characters(source, new_tree.meta.end_pos, len(source)))
-    )
-    final_tree = Tree(new_tree.data, final_children, final_meta)
-    return final_tree
-
-
 @install_grammar
 class ThetaRecordParser(RecordParser):
     grammar_filename = 'theta_record.lark'
@@ -161,7 +61,7 @@ class ThetaRecordParser(RecordParser):
     non_empty = [
         {'comment': (1, 'COMMENT')},
     ]
-    post_process = [InitOrLow(), insert_ignored]
+    post_process = [InitOrLow(), with_ignored_tokens]
 
 
 @install_grammar
@@ -173,7 +73,7 @@ class OmegaRecordParser(RecordParser):
     non_empty = [
         {'comment': (1, 'COMMENT')},
     ]
-    post_process = [insert_ignored]
+    post_process = [with_ignored_tokens]
 
 
 @install_grammar
